Started a database model file : models.py
Edited call_database.py to work with environment variables
